model09

+++ Module_00/ex00/megaphone.cpp
@@ -18,10 +18,10 @@ int main(int argv, char **argc)
 	std::string arg;
 	if (argv == 1)
 		return (0);
-	for (size_t i = 1; i < argv; i++)
+	for (int i = 1; i < argv; i++)
 	{
 		arg = std::string(argc[i]);
-		for (int j = 0; j < arg.length(); j++)
+		for (size_t j = 0; j < arg.length(); j++)
 			std::cout << (char) std::toupper(arg[j]);
 	}
 	std::cout << "\n";
